requirements.txt
# CodeRetrievr 后端依赖

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库连接
pymilvus==2.3.4
neo4j==5.14.1
sqlalchemy==2.0.23

# 代码嵌入和AI
sentence-transformers==2.2.2
transformers==4.35.0
# PyTorch安装说明：
# - 如果有NVIDIA GPU且想使用CUDA加速，请先安装CUDA版本的PyTorch：
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# - 如果只有CPU或想使用CPU版本，可以安装CPU版本：
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# - 或者使用conda安装（推荐）：
#   conda install pytorch pytorch-cuda=11.8 -c pytorch -c nvidia
# 注意：不要直接pip install torch，这会安装CPU版本
# torch  # 注释掉，需要根据上述说明手动安装
openai==1.3.0

# 代码解析
# 注意：tree-sitter-python 新版本需要 Python 3.10+
# 如果使用 Python 3.9，可能需要升级到 Python 3.10 或使用替代方案
tree-sitter
tree-sitter-python  # pip会自动选择兼容版本，Python 3.9可能需要升级到3.10
tree-sitter-java
tree-sitter-cpp

# GitHub API
PyGithub==1.59.1
requests==2.31.0

# 数据处理
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1  # 进度条

# 工具库
python-dotenv==1.0.0
httpx==0.25.1
aiofiles==23.2.1

# 日志和监控
structlog==23.2.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# 代码质量
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

